backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.0
numpy==1.24.0
pandas==2.0.3
duckduckgo-search>=7.0.0
chromadb>=0.4.22
ollama>=0.4.0
pymongo==4.6.0
yfinance==1.1.0
